fix: Resolve all small bugs
