+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "zero",
-  "version": "0.0.1",
+  "version": "1.0",
   "private": true,
   "scripts": {
     "android": "react-native run-android",
