+++ src/components/molecules/DebtEntry.tsx
@@ -105,142 +105,149 @@ const DebtEntry: React.FC<DebtEntryProps> = ({buttonText, route}) => {
   };
 
   return (
-    <PrimaryView colors={colors}>
-      <View style={mainStyles.headerContainer}>
-        <AppHeader
-          onPress={goBack}
-          colors={colors}
-          text={`${buttonText} Debt ◦ ${debtorName}`}
-        />
-      </View>
-      <View
-        style={{
-          flexDirection: 'row',
-          width: '100%',
-          alignItems: 'center',
-          justifyContent: 'center',
-          marginBottom: 15,
-        }}>
-        <TouchableOpacity
-          onPress={() => setDebtsType('Borrow')}
-          style={[
-            debtsStyles.categoryContainer,
-            {
-              backgroundColor:
-                debtsType === 'Borrow'
-                  ? colors.accentOrange
-                  : colors.secondaryAccent,
-              borderColor: colors.secondaryContainerColor,
-              width: '48.5%',
-              height: 50,
-            },
-          ]}>
-          <PrimaryText
-            style={{
-              color:
-                debtsType === 'Borrow' ? colors.buttonText : colors.primaryText,
-              fontSize: 13,
-              fontFamily: 'FiraCode-SemiBold',
-            }}>
-            Borrowing from
-          </PrimaryText>
-          <PrimaryText
-            style={{
-              color:
-                debtsType === 'Borrow' ? colors.buttonText : colors.primaryText,
-              fontSize: 13,
-              fontFamily: 'FiraCode-SemiBold',
-            }}>
-            {debtorName}
-          </PrimaryText>
-        </TouchableOpacity>
-        <TouchableOpacity
-          onPress={() => setDebtsType('Lend')}
-          style={[
-            debtsStyles.categoryContainer,
-            {
-              backgroundColor:
-                debtsType === 'Lend'
-                  ? colors.accentGreen
-                  : colors.secondaryAccent,
-              borderColor: colors.secondaryContainerColor,
-              width: '48.5%',
-              height: 50,
-            },
-          ]}>
-          <PrimaryText
-            style={{
-              color:
-                debtsType === 'Lend' ? colors.buttonText : colors.primaryText,
-              fontSize: 13,
-              fontFamily: 'FiraCode-SemiBold',
-            }}>
-            Lending to
-          </PrimaryText>
-          <PrimaryText
-            style={{
-              color:
-                debtsType === 'Lend' ? colors.buttonText : colors.primaryText,
-              fontSize: 13,
-              fontFamily: 'FiraCode-SemiBold',
-            }}>
-            {debtorName}
-          </PrimaryText>
-        </TouchableOpacity>
-      </View>
-
-      <CustomInput
-        colors={colors}
-        input={debtName}
-        setInput={setDebtName}
-        placeholder="eg. tea"
-        label={`${debtsType} Title`}
-        schema={expenseSchema}
-      />
-      <PrimaryText style={{marginBottom: 5}}>{debtsType} Amount</PrimaryText>
-
-      <View
-        style={[
-          textInputStyles.textInputContainer,
-          {
-            borderColor: colors.secondaryContainerColor,
-            backgroundColor: colors.secondaryAccent,
-            marginBottom: debtAmountError.length > 0 ? 5 : 15,
-          },
-        ]}>
-        <PrimaryText style={{fontSize: 15}}>{currencySymbol}</PrimaryText>
-        <TextInput
-          style={[
-            textInputStyles.textInputWithIcon,
-            {
-              color: colors.primaryText,
-            },
-          ]}
-          value={debtAmount}
-          onChangeText={setDebtAmount}
-          placeholder={'eg. 200'}
-          placeholderTextColor={colors.secondaryText}
-          keyboardType="numeric"
-        />
-      </View>
-      {debtAmountError.length > 0 && (
-        <View style={{marginBottom: 10}}>
-          {debtAmountError.map(error => (
-            <View key={error.message}>
-              <PrimaryText style={{color: colors.accentRed, fontSize: 12}}>
-                {error.message}
-              </PrimaryText>
-            </View>
-          ))}
+    <PrimaryView colors={colors} style={{justifyContent: 'space-between'}}>
+      <View>
+        <View style={mainStyles.headerContainer}>
+          <AppHeader
+            onPress={goBack}
+            colors={colors}
+            text={`${buttonText} Debt ◦ ${debtorName}`}
+          />
         </View>
-      )}
-      <DatePicker
-        setShowDatePicker={setShowDatePicker}
-        createdAt={createdAt}
-        showDatePicker={showDatePicker}
-        setCreatedAt={setCreatedAt}
-      />
-      <View style={isValid ? {marginTop: '93%'} : {marginTop: '80%'}}>
+        <View
+          style={{
+            flexDirection: 'row',
+            width: '100%',
+            alignItems: 'center',
+            justifyContent: 'center',
+            marginBottom: 15,
+          }}>
+          <TouchableOpacity
+            onPress={() => setDebtsType('Borrow')}
+            style={[
+              debtsStyles.categoryContainer,
+              {
+                backgroundColor:
+                  debtsType === 'Borrow'
+                    ? colors.accentOrange
+                    : colors.secondaryAccent,
+                borderColor: colors.secondaryContainerColor,
+                width: '48.5%',
+                height: 50,
+              },
+            ]}>
+            <PrimaryText
+              style={{
+                color:
+                  debtsType === 'Borrow'
+                    ? colors.buttonText
+                    : colors.primaryText,
+                fontSize: 13,
+                fontFamily: 'FiraCode-SemiBold',
+              }}>
+              Borrowing from
+            </PrimaryText>
+            <PrimaryText
+              style={{
+                color:
+                  debtsType === 'Borrow'
+                    ? colors.buttonText
+                    : colors.primaryText,
+                fontSize: 13,
+                fontFamily: 'FiraCode-SemiBold',
+              }}>
+              {debtorName}
+            </PrimaryText>
+          </TouchableOpacity>
+          <TouchableOpacity
+            onPress={() => setDebtsType('Lend')}
+            style={[
+              debtsStyles.categoryContainer,
+              {
+                backgroundColor:
+                  debtsType === 'Lend'
+                    ? colors.accentGreen
+                    : colors.secondaryAccent,
+                borderColor: colors.secondaryContainerColor,
+                width: '48.5%',
+                height: 50,
+              },
+            ]}>
+            <PrimaryText
+              style={{
+                color:
+                  debtsType === 'Lend' ? colors.buttonText : colors.primaryText,
+                fontSize: 13,
+                fontFamily: 'FiraCode-SemiBold',
+              }}>
+              Lending to
+            </PrimaryText>
+            <PrimaryText
+              style={{
+                color:
+                  debtsType === 'Lend' ? colors.buttonText : colors.primaryText,
+                fontSize: 13,
+                fontFamily: 'FiraCode-SemiBold',
+              }}>
+              {debtorName}
+            </PrimaryText>
+          </TouchableOpacity>
+        </View>
+
+        <CustomInput
+          colors={colors}
+          input={debtName}
+          setInput={setDebtName}
+          placeholder="eg. tea"
+          label={`${debtsType} Title`}
+          schema={expenseSchema}
+        />
+        <PrimaryText style={{marginBottom: 5}}>{debtsType} Amount</PrimaryText>
+
+        <View
+          style={[
+            textInputStyles.textInputContainer,
+            {
+              borderColor: colors.secondaryContainerColor,
+              backgroundColor: colors.secondaryAccent,
+              marginBottom: debtAmountError.length > 0 ? 5 : 15,
+            },
+          ]}>
+          <PrimaryText style={{fontSize: 15}}>{currencySymbol}</PrimaryText>
+          <TextInput
+            style={[
+              textInputStyles.textInputWithIcon,
+              {
+                color: colors.primaryText,
+              },
+            ]}
+            value={debtAmount}
+            onChangeText={setDebtAmount}
+            placeholder={'eg. 200'}
+            placeholderTextColor={colors.secondaryText}
+            keyboardType="numeric"
+          />
+        </View>
+        {debtAmountError.length > 0 && (
+          <View style={{marginBottom: 10}}>
+            {debtAmountError.map(error => (
+              <View key={error.message}>
+                <PrimaryText style={{color: colors.accentRed, fontSize: 12}}>
+                  {error.message}
+                </PrimaryText>
+              </View>
+            ))}
+          </View>
+        )}
+        <DatePicker
+          setShowDatePicker={setShowDatePicker}
+          createdAt={createdAt}
+          showDatePicker={showDatePicker}
+          setCreatedAt={setCreatedAt}
+        />
+      </View>
+
+      <View style={{marginBottom: '10%'}}>
         <PrimaryButton
           onPress={isAddButton ? handleAddDebt : handleUpdateDebt}
           colors={colors}
