+++ src/components/molecules/DebtorEntry.tsx
@@ -16,7 +16,7 @@ import {createDebtor, updateDebtorById} from '../../services/DebtorService';
 import {FETCH_ALL_DEBTOR_DATA} from '../../redux/actionTypes';
 import debtCategories from '../../../assets/jsons/defaultDebtAccounts.json';
 import {nameSchema} from '../../utils/validationSchema';
-import { getDebtRequest } from '../../redux/slice/debtDataSlice';
+import {getDebtRequest} from '../../redux/slice/debtDataSlice';
 
 interface DebtorEntryProps {
   type: string;
@@ -84,31 +84,34 @@ const DebtorEntry: React.FC<DebtorEntryProps> = ({type, route}) => {
   };
 
   return (
-    <PrimaryView colors={colors}>
-      <View style={mainStyles.headerContainer}>
-        <AppHeader onPress={goBack} colors={colors} text={`${type} Debtor`} />
-      </View>
-      <PrimaryText style={{marginBottom: 5}}>Select Debt Category</PrimaryText>
+    <PrimaryView colors={colors} style={{justifyContent: 'space-between'}}>
+      <View>
+        <View style={mainStyles.headerContainer}>
+          <AppHeader onPress={goBack} colors={colors} text={`${type} Debtor`} />
+        </View>
+        <PrimaryText style={{marginBottom: 5}}>
+          Select Debt Category
+        </PrimaryText>
 
-      <View style={{marginBottom: 10}}>
-        <CategoryContainer
-          categories={debtCategories}
+        <View style={{marginBottom: 10}}>
+          <CategoryContainer
+            categories={debtCategories}
+            colors={colors}
+            toggleCategorySelection={toggleCategorySelection}
+            selectedCategories={selectedCategories}
+          />
+        </View>
+
+        <CustomInput
           colors={colors}
-          toggleCategorySelection={toggleCategorySelection}
-          selectedCategories={selectedCategories}
+          input={debtorTitle}
+          setInput={setDebtorTitle}
+          placeholder="eg. John Doe or Axis"
+          label="Debtor Name"
+          schema={nameSchema}
         />
       </View>
-
-      <CustomInput
-        colors={colors}
-        input={debtorTitle}
-        setInput={setDebtorTitle}
-        placeholder="eg. John Doe or Axis"
-        label="Debtor Name"
-        schema={nameSchema}
-      />
-
-      <View style={isValid ? {marginTop: '115%'} : {marginTop: '105%'}}>
+      <View style={{marginBottom: '10%'}}>
         <PrimaryButton
           onPress={isAddButton ? handleAddDebtor : handleUpdateDebtor}
           colors={colors}
