+++ src/screens/PersonalizeScreen/index.tsx
@@ -13,48 +13,55 @@ const PersonalizeScreen = () => {
   const isValid = nameSchema.safeParse(name).success;
 
   return (
-    <PrimaryView colors={colors}>
-      <TouchableOpacity style={styles.skipButtonContainer} onPress={handleSkip}>
-        <PrimaryText style={{color: colors.accentGreen, fontSize: 12}}>
-          skip
-        </PrimaryText>
-      </TouchableOpacity>
+    <PrimaryView colors={colors} style={{justifyContent: 'space-between'}}>
+      <View>
+        <TouchableOpacity
+          style={styles.skipButtonContainer}
+          onPress={handleSkip}>
+          <PrimaryText style={{color: colors.accentGreen, fontSize: 12}}>
+            skip
+          </PrimaryText>
+        </TouchableOpacity>
 
-      <View style={styles.titleTextContainer}>
-        <PrimaryText style={{fontSize: 24}}>Let's Personlize your</PrimaryText>
-        <PrimaryText style={{fontSize: 24}}>experience</PrimaryText>
+        <View style={styles.titleTextContainer}>
+          <PrimaryText style={{fontSize: 24}}>
+            Let's Personlize your
+          </PrimaryText>
+          <PrimaryText style={{fontSize: 24}}>experience</PrimaryText>
+        </View>
+
+        <View style={styles.subtitleTextContainer}>
+          <PrimaryText style={{color: colors.accentGreen, fontSize: 15}}>
+            Hi, It's zero! What Do Your Friends
+          </PrimaryText>
+          <PrimaryText style={{color: colors.accentGreen, fontSize: 15}}>
+            Call You?
+          </PrimaryText>
+        </View>
+
+        <View
+          style={[
+            styles.textInputContainer,
+            isValid ? {marginBottom: '100%'} : null,
+          ]}>
+          <CustomInput
+            input={name}
+            label={'Name'}
+            colors={colors}
+            placeholder={'eg. Indranil Bhuin'}
+            setInput={setName}
+            schema={nameSchema}
+          />
+        </View>
       </View>
-
-      <View style={styles.subtitleTextContainer}>
-        <PrimaryText style={{color: colors.accentGreen, fontSize: 15}}>
-          Hi, It's zero! What Do Your Friends
-        </PrimaryText>
-        <PrimaryText style={{color: colors.accentGreen, fontSize: 15}}>
-          Call You?
-        </PrimaryText>
-      </View>
-
-      <View
-        style={[
-          styles.textInputContainer,
-          isValid ? {marginBottom: '100%'} : null,
-        ]}>
-        <CustomInput
-          input={name}
-          label={'Name'}
+      <View style={{marginBottom: '10%'}}>
+        <PrimaryButton
+          onPress={handleSubmit}
           colors={colors}
-          placeholder={'eg. Indranil Bhuin'}
-          setInput={setName}
-          schema={nameSchema}
+          buttonTitle={'Continue'}
+          disabled={!isValid}
         />
       </View>
-
-      <PrimaryButton
-        onPress={handleSubmit}
-        colors={colors}
-        buttonTitle={'Continue'}
-        disabled={!isValid}
-      />
     </PrimaryView>
   );
 };
